% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_colData.R
\name{cbind_colData}
\alias{cbind_colData}
\title{Joins dds colData with rowname matching}
\usage{
cbind_colData(colData_1, colData_2)
}
\arguments{
\item{colData_1}{colData to be joined TO}

\item{colData_2}{coldata to join}
}
\value{
joined colData
}
\description{
Joins dds colData with rowname matching
}
