% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_heatmap.R
\name{quick_heatmap}
\alias{quick_heatmap}
\title{Quickly make group-stratified heatmaps}
\usage{
quick_heatmap(dds, genes = c(), stratify_variable = c())
}
\arguments{
\item{dds}{A DDS object with the \code{assay(dds, 2)} containing normalized counts.}

\item{genes}{A list containing hgnc symbols.}

\item{stratify_variable}{A character list of colnames from \code{colData(dds)} on which to stratify}
}
\value{
A pheatmap
}
\description{
Quickly make group-stratified heatmaps
}
